feat(takeout): implement new features and components for takeout page
