+++ code/tamagui.dev/app.css
@@ -842,3 +842,56 @@ p {
     animation: none;
   }
 }
+
+/* Orbit animation for takeout page */
+.orbit-container {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  display: grid;
+  place-items: center;
+  pointer-events: none;
+  overflow: visible;
+}
+
+.orbit-planet {
+  --offset: 520px;
+  grid-area: 1 / 1;
+  width: 60px;
+  height: 60px;
+  will-change: transform;
+  contain: layout style;
+  animation: orbitSpin 20s linear infinite, orbitZ 20s step-end infinite;
+}
+
+.orbit-planet img {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+}
+
+@keyframes orbitSpin {
+  from {
+    transform: rotate(0deg) translateX(var(--offset)) rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg) translateX(var(--offset)) rotate(-360deg);
+  }
+}
+
+@keyframes orbitZ {
+  0%,
+  24.99% {
+    z-index: 1;
+  }
+  25%,
+  74.99% {
+    z-index: -1;
+  }
+  75%,
+  100% {
+    z-index: 1;
+  }
+}
